+++ src/lib.rs
@@ -17,24 +17,12 @@
 #![deny(missing_docs)]
 #![warn(missing_debug_implementations)]
 
-mod arc_list;
-mod atomic_waker;
 #[cfg(not(all(target_arch = "wasm32", feature = "wasm-bindgen")))]
-mod delay;
+mod native;
 #[cfg(all(target_arch = "wasm32", feature = "wasm-bindgen"))]
-mod delay_wasm;
-mod global;
-mod heap;
-mod heap_timer;
-mod timer;
-
-use arc_list::{ArcList, Node};
-use atomic_waker::AtomicWaker;
-use heap::{Heap, Slot};
-use heap_timer::HeapTimer;
-use timer::{ScheduledTimer, Timer, TimerHandle};
+mod wasm;
 
 #[cfg(not(all(target_arch = "wasm32", feature = "wasm-bindgen")))]
-pub use self::delay::Delay;
+pub use self::native::Delay;
 #[cfg(all(target_arch = "wasm32", feature = "wasm-bindgen"))]
-pub use self::delay_wasm::Delay;
+pub use self::wasm::Delay;
