+++ src/native/heap.rs
