+++ src/wasm.rs
